src/requirements.txt
Flask==1.1.1
Flask-WTF==0.14.3
Flask-SQLAlchemy==2.4.1
Flask-Migrate==2.5.3
psycopg2-binary==2.8.4
Flask-Admin==1.5.5
Flask-BabelEx==0.9.4
Flask-Mail==0.9.1
validate_email==1.3